Refactor linear solver into modular components with clean architecture
